+++ src/modules/titulacion/dto/myData.ts
@@ -1,0 +1,75 @@
+
+export const myData = [
+    {
+      numero: 1,
+      escuela: "Escuela 1",
+      carrera: "Carrera 1",
+      activa: true,
+    },
+    {
+      numero: 2,
+      escuela: "Escuela 2",
+      carrera: "Carrera 2",
+      activa: false,
+    },
+    {
+      numero: 3,
+      escuela: "Escuela 3",
+      carrera: "Carrera 3",
+      activa: true,
+    },
+    {
+      numero: 4,
+      escuela: "Escuela 1",
+      carrera: "Carrera 1",
+      activa: true,
+    },
+    {
+      numero: 5,
+      escuela: "Escuela 2",
+      carrera: "Carrera 2",
+      activa: false,
+    },
+    
+    // Agrega más datos según sea necesario
+  ];
+  
+  
+export const dataTitulo=[
+  {
+    numero: 1,
+    cedula: "1002003001",
+    nombre: "juan",
+    modalidad: "Trabajo de titulación",
+    titulo: "ingeniero",
+    malla: "2009",
+    defensa: "2023-10-02",
+    informe: "boton",
+    respaldo: "circulo",
+    existencia: "vacio",
+  },
+  {
+    numero: 2,
+    cedula: "1002003001",
+    nombre: "juan",
+    modalidad: "Trabajo de titulación",
+    titulo: "ingeniero",
+    malla: "2009",
+    defensa: "2023-10-02",
+    informe: "boton",
+    respaldo: "circulo",
+    existencia: "vacio",
+  },
+  {
+    numero: 3,
+    cedula: "1002003001",
+    nombre: "juan",
+    modalidad: "Trabajo de titulación",
+    titulo: "ingeniero",
+    malla: "2009",
+    defensa: "2023-10-02",
+    informe: "boton",
+    respaldo: "circulo",
+    existencia: "vacio",
+  },
+]
